# Re-implementing gather op